Elta/adventure works iss

+++ src/adventure_works/models/prp/adventure_works_full/sales/SalesOrderDetail_prp.sql
@@ -11,9 +11,9 @@ prepared_source as (
         CAST(carriertrackingnumber as string) as numero_suivi_transporteur,
         CAST(orderqty as int64) as quantite_commandee,
         CAST(productid as int64) as id_produit,
-        CAST(specialofferid as int64) as id_offre_promotionnelle,
-        CAST(unitprice as numeric) as prix_unitaire,
-        CAST(unitpricediscount as numeric) as remise,
+        CAST(specialofferid as float64) as id_offre_promotionnelle,
+        CAST(unitprice as float64) as prix_unitaire,
+        CAST(unitpricediscount as float64) as remise,
         CAST(linetotal as bignumeric) as total_ligne,
         CAST(rowguid as string) as id_unique,
         CAST(modifieddate as datetime) as date_modification
